+++ src/master/masterconn.cc
@@ -636,7 +636,8 @@ void MasterConn::downloadNext() {
 		if (fileSize - downloadOffset > kMetadataDownloadBlocksize) {
 			put32bit(&ptr, kMetadataDownloadBlocksize);
 		} else {
-			put32bit(&ptr, fileSize - downloadOffset);
+			// TODO(Guillex): Possible truncation if the file size is too big
+			put32bit(&ptr, static_cast<uint32_t>(fileSize - downloadOffset));
 		}
 	}
 }
